refactor(ui): extract custom option item
